+++ open_city_profile/graphene.py
@@ -5,6 +5,15 @@ from django.forms import MultipleChoiceField
 from django_filters import MultipleChoiceFilter
 from graphene_django.forms.converter import convert_form_field
 
+from profiles.loaders import (
+    AddressesByProfileIdLoader,
+    EmailsByProfileIdLoader,
+    PhonesByProfileIdLoader,
+    PrimaryAddressForProfileLoader,
+    PrimaryEmailForProfileLoader,
+    PrimaryPhoneForProfileLoader,
+)
+
 
 class JWTMiddleware:
     def resolve(self, next, root, info, **kwargs):
@@ -35,3 +44,29 @@ def convert_form_field_to_uuid_list(field):
 
 class UUIDMultipleChoiceFilter(MultipleChoiceFilter):
     field_class = UUIDMultipleChoiceField
+
+
+_LOADERS = {
+    "addresses_by_profile_id_loader": AddressesByProfileIdLoader,
+    "emails_by_profile_id_loader": EmailsByProfileIdLoader,
+    "phones_by_profile_id_loader": PhonesByProfileIdLoader,
+    "primary_address_for_profile_loader": PrimaryAddressForProfileLoader,
+    "primary_email_for_profile_loader": PrimaryEmailForProfileLoader,
+    "primary_phone_for_profile_loader": PrimaryPhoneForProfileLoader,
+}
+
+
+class GQLDataLoaders:
+    def __init__(self):
+        self.cached_loaders = False
+
+    def resolve(self, next, root, info, **kwargs):
+        context = info.context
+
+        if not self.cached_loaders:
+            for loader_name, loader_class in _LOADERS.items():
+                setattr(context, loader_name, loader_class())
+
+            self.cached_loaders = True
+
+        return next(root, info, **kwargs)
